+++ libraries/tock-rt0/src/lib.rs
@@ -12,9 +12,9 @@
 /// - `end_data_ram`:     The first address after the end of the data section
 ///                       in RAM.
 pub unsafe fn init_data(
-    mut start_data_flash: *mut u32,
-    mut start_data_ram: *mut u32,
-    end_data_ram: *mut u32,
+    mut start_data_flash: *mut usize,
+    mut start_data_ram: *mut usize,
+    end_data_ram: *mut usize,
 ) {
     while start_data_ram < end_data_ram {
         start_data_ram.write(start_data_flash.read());
@@ -25,7 +25,7 @@ pub unsafe fn init_data(
 
 /// Sets non-initialized data in RAM to 0.
 /// This is used to clear the BSS section on initial bootup.
-pub unsafe fn zero_bss(mut start_bss: *mut u32, end_bss: *mut u32) {
+pub unsafe fn zero_bss(mut start_bss: *mut usize, end_bss: *mut usize) {
     while start_bss < end_bss {
         // `volatile` to make sure it doesn't get optimized out
         start_bss.write_volatile(0);
